move expansion 2/?
give all the moves names and descriptions that currently match with egg's engine.
update some screens on fairy type, the party menu as it currently is.
no more using task data array to store ball sprite id's, now just reserving 6 more bytes for it.
reintegrate ball printing back in with the icons so that all of the battle screens get it without too much editing