rtc is now based on play time
the time also displays on the continue menu, the start menu, etc.
there needs to be a better way to do the start menu so that we definitely don't blend the palettes there.
add timeblendaffectedpalettes to the overworld main func.  it would be better if we could track all this.
made sure to load the things correctly as well.  need to track down where the palettes for secondary tilesets are loaded on transition
add playtimeyears and playtimedays to the saveblock2 structure
document some more bugs